SCRUM-35: implement react router and removed header to the Head.js
