+++ frontend/package.json
@@ -13,6 +13,7 @@
     "react": "^19.0.0",
     "react-bootstrap": "^2.10.8",
     "react-dom": "^19.0.0",
+    "react-router-dom": "^7.1.5",
     "react-scripts": "^5.0.1",
     "web-vitals": "^4.2.4"
   },
